Sometimes C++ is weird.
The GCC running on the CI complains that it can't deduce the return type of these functions (error: "auto" function requires a trailing return type). Google says, that the solution is to use -> decltype(...), but then Visual Studio complains that it does not understand that.
The solution: spell out the actual return types manually. Ugly!